Add initial smtp implementation
It currently only supports a plaintext body and very little error
and status reporting (still to be done)

This adds a SubmissionManager and MessageBuilder components.
MessageBuilder is resposible for the collecting of data and creation
of a valid QMailMessage which SubmissionManager does some validation
on before passing it on the ClientService.

ClientService will first store the message in the local outbox and
then send the message. We still need to add some kind of
timer/check interval to send pending messages using the sendAllPending
slot. As there may have been no network so we should be timely in
getting it submitted when the opportunity arrises.