Add Forward message
For now we only support forward inline if it's just a text/plain or multipart/alternative with a text/plain part. In some cases we will also convert a text/html part to text/plain and use that inline. For all other multipart/* parts we wrap them in a message/rfc822 and attach them as a .eml attachment.