Add notification uri handler
Notifictions urls can be in 3 forms

dekko://notify/accountId - opens the UI to the accounts folder list
dekko://notify/accountId/folderName - opens to account folder
dekko://notify/accountId/folderName/messageUid - opens message in folder

The first isn't so useful but the second can be used when a notification
is a grouped notification consisting of more than one message and
opening the folder they belong to is the only sane thing to do. The last
is for single message notifications

We still need to add ViewAction handlers for the notifications. Coming
up next ...