Remove duplicate code in atomic settings subclasses
All subclasses now need to do is set a settings key in the constructor. SettingsObjectBase will keep a pointer map to the settingsfilebase and re-use it for all instances of settings key. The settings key is also used to name the settings file and settings lock file. i.e settingsfile.json & settingsfile.json.lock