Add start of settings policies
A Policy is related set of configurations/settings that apply to a single account. You can either create an instance of each policy using the account id or there is a PolicyManager singleton to have a single point of entry to each policy. Which is nicer in the c++ world :-)

The remaining policies and the actual configurations will be coming in the next commit